refactor(optimization/cached_expression): CachedExpression
Previously, it was part of nlopt_optimizer.h. Now it is in a separate file.
